# crazygames-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/game": {
            "get": {
                "description": "Get all games",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Games"
                ],
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "page_number",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.GamesResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new game",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Games"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "game_title",
                        "name": "game_title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "description",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "developer",
                        "name": "developer",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "category_id",
                        "name": "category_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "release_date",
                        "name": "release_date",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "thumbnail",
                        "name": "thumbnail",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "technology",
                        "name": "technology",
                        "in": "formData"
                    },
                    {
                        "type": "number",
                        "description": "rating",
                        "name": "rating",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "hover_video",
                        "name": "hover_video",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "game_url",
                        "name": "game_url",
                        "in": "formData"
                    },
                    {
                        "type": "number",
                        "description": "play_count",
                        "name": "play_count",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.Game"
                        }
                    }
                }
            }
        },
        "/game/category/{id}": {
            "get": {
                "description": "Get a game by category id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Games"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Game"
                        }
                    }
                }
            }
        },
        "/game/{id}": {
            "get": {
                "description": "List games by category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Games"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "category_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/entities.Game"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update a game",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Games"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "game_title",
                        "name": "game_title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "description",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "developer",
                        "name": "developer",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "category_id",
                        "name": "category_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "release_date",
                        "name": "release_date",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "thumbnail",
                        "name": "thumbnail",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "technology",
                        "name": "technology",
                        "in": "formData"
                    },
                    {
                        "type": "number",
                        "description": "rating",
                        "name": "rating",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "hover_video",
                        "name": "hover_video",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "game_url",
                        "name": "game_url",
                        "in": "formData"
                    },
                    {
                        "type": "number",
                        "description": "play_count",
                        "name": "play_count",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Game"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a game",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Games"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Game"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entities.Category": {
            "type": "object",
            "properties": {
                "categoryName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "game": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Game"
                    }
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isMenu": {
                    "type": "boolean"
                },
                "path": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "entities.Game": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Category"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "developer": {
                    "type": "string"
                },
                "gameTitle": {
                    "type": "string"
                },
                "gameURL": {
                    "type": "string"
                },
                "hoverVideoUrl": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "playCount": {
                    "type": "integer"
                },
                "rating": {
                    "type": "number"
                },
                "releaseDate": {
                    "type": "string"
                },
                "technology": {
                    "type": "string"
                },
                "thumbnailURL": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "response.GamesResponse": {
            "type": "object",
            "properties": {
                "games": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Game"
                    }
                },
                "pageNumber": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}